requirements.txt
requests
pandas
python-dotenv
openai
openpyxl
mysql-connector-python
boto3
tqdm 